e2e: Support upgrading local clusters
Allow upgrade tests to be run on local clusters by allowing callers to
provide a script that will upgrade the cluster. Since every local
cluster will be different we can't provide a default script, so the
default is to throw an error - the current behavior when running upgrade
tests on a local cluster.

This will allow for running e2e tests against, e.g., managed kubernetes
providers that support upgrades via an API.